Snapshot improvements
-  error handling
- consolidate a bunch of utils
- move some utils to lib.
