+++ src/components/v2/V2Project/V2ProjectSettings/V2ProjectSettings.tsx
@@ -25,10 +25,9 @@ export const V2SettingsPageKeyTitleMap: { [k in V2SettingsPageKey]: string } = {
   payouts: t`Payouts`,
   reservedtokens: t`Reserved Token Allocation`,
   v1tokenmigration: t`V1 Token Migration`,
-  venft: t`VeNFT Governance`,
   transferownership: t`Transfer Ownership`,
   archiveproject: t`Archive Project`,
-  tokenweightedvoting: t`Token-Weighted Voting`,
+  governance: t`Governance`,
 }
 
 function menuItem(
@@ -73,13 +72,10 @@ const items: MenuItem[] = [
     [
       menuItem('Transfer ownership', 'transferownership'),
       menuItem('Archive project', 'archiveproject'),
-      featureFlagEnabled(FEATURE_FLAGS.VENFT)
-        ? menuItem('veNFT governance', 'venft')
-        : null,
+      menuItem('Governance', 'governance'),
       featureFlagEnabled(FEATURE_FLAGS.V1_TOKEN_SWAP)
         ? menuItem('V1 token migration', 'v1tokenmigration')
         : null,
-      menuItem('Token-Weighted Voting', 'tokenweightedvoting'),
     ],
     'group',
   ),
