+++ src/components/v2/V2Project/V2ProjectSettings/pages/GovernanceSettingsPage/index.ts
@@ -1,0 +1,1 @@
+export { GovernanceSettingsPage } from './GovernanceSettingsPage'
