+++ src/models/menu-keys.ts
@@ -8,8 +8,7 @@ export type V2SettingsPageKey =
   | 'reconfigurefc'
   | 'payouts'
   | 'reservedtokens'
-  | 'v1tokenmigration'
-  | 'venft'
   | 'transferownership'
   | 'archiveproject'
-  | 'tokenweightedvoting'
+  | 'governance'
+  | 'v1tokenmigration'
