+++ src/pages/p/[handle].page.tsx
@@ -29,13 +29,13 @@ import useQueuedPayoutModsOfProject from 'hooks/v1/contractReader/QueuedPayoutMo
 import useQueuedTicketModsOfProject from 'hooks/v1/contractReader/QueuedTicketModsOfProject'
 import useTerminalOfProject from 'hooks/v1/contractReader/TerminalOfProject'
 import useTokenAddressOfProject from 'hooks/v1/contractReader/TokenAddressOfProject'
+import { paginateDepleteProjectsQueryCall } from 'lib/apollo'
 import { ProjectMetadataV4 } from 'models/project-metadata'
 import { V1CurrencyOption } from 'models/v1/currencyOption'
 import { GetStaticPaths, GetStaticProps, InferGetStaticPropsType } from 'next'
 import { useRouter } from 'next/router'
 import { V1CurrencyProvider } from 'providers/v1/V1CurrencyProvider'
 import { useMemo } from 'react'
-import { paginateDepleteProjectsQueryCall } from 'utils/apollo'
 import { findProjectMetadata } from 'utils/server'
 import { V1CurrencyName } from 'utils/v1/currency'
 import { getTerminalName, getTerminalVersion } from 'utils/v1/terminals'
